Separate console code
Console code is in fact very system-specific, let's keep it in a separate for
easy replacement
