+++ console.h
@@ -1,0 +1,2 @@
+void tty_init(void);
+void net_init(unsigned short port);
