+++ mux.c
@@ -1,30 +1,11 @@
 #include <errno.h>
-#include <fcntl.h>
-#include <signal.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <termios.h>
 #include <unistd.h>
-#include <sys/select.h>
-#include <sys/socket.h>
-#include <arpa/inet.h>
 
 #include "centurion.h"
 #include "cpu6.h"
 #include "mux.h"
-
-static struct termios saved_term, term;
-
-static void cleanup(int sig)
-{
-	tcsetattr(0, TCSADRAIN, &saved_term);
-	emulator_done = 1;
-}
-
-static void exit_cleanup(void)
-{
-	tcsetattr(0, TCSADRAIN, &saved_term);
-}
 
 static struct MuxUnit mux[NUM_MUX_UNITS];
 static char ipl_request = -1; // Let's suppose -1 = disabled
@@ -42,61 +23,10 @@ void mux_init(void)
         }
 }
 
-void tty_init(void)
+void mux_attach(unsigned unit, int in_fd, int out_fd)
 {
-	if (tcgetattr(0, &term) == 0) {
-		saved_term = term;
-		atexit(exit_cleanup);
-		signal(SIGINT, cleanup);
-		signal(SIGQUIT, cleanup);
-		signal(SIGPIPE, cleanup);
-		term.c_iflag &= ~ICRNL;
-		term.c_lflag &= ~(ICANON | ECHO);
-		term.c_cc[VMIN] = 0;
-		term.c_cc[VTIME] = 1;
-		term.c_cc[VINTR] = 0;
-		term.c_cc[VSUSP] = 0;
-		term.c_cc[VSTOP] = 0;
-		tcsetattr(0, TCSADRAIN, &term);
-	}
-
-        mux[0].in_fd = STDIN_FILENO;
-        mux[0].out_fd = STDOUT_FILENO;
-}
-
-void net_init(unsigned short port)
-{
-	struct sockaddr_in sin;
-        int sock_fd, io_fd;
-
-	sock_fd = socket(AF_INET, SOCK_STREAM, 0);
-	if (sock_fd == -1) {
-		perror("socket");
-		exit(1);
-	}
-	sin.sin_family = AF_INET;
-	sin.sin_addr.s_addr = htonl(0x7F000001);
-	sin.sin_port = htons(port);
-	if (bind(sock_fd, (struct sockaddr *) &sin, sizeof(sin)) == -1) {
-		perror("bind");
-		exit(1);
-	}
-	/* TODO set reuseaddr */
-	listen(sock_fd, 1);
-
-	printf("[Waiting terminal connection...]\n");
-	fflush(stdout);
-
-	io_fd = accept(sock_fd, NULL, NULL);
-	if (io_fd == -1) {
-		perror("accept");
-		exit(1);
-	}
-	close(sock_fd);
-	fcntl(io_fd, F_SETFL, FNDELAY);
-
-        mux[0].in_fd = io_fd;
-	mux[0].out_fd = io_fd;
+	mux[unit].in_fd = in_fd;
+	mux[unit].out_fd = out_fd;
 }
 
 static int select_wrapper(int maxfd, fd_set* i, fd_set* o) {
